requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124

# macOS (CPU/MPS)
torch==2.6.0 ; platform_system == "Darwin"
torchvision==0.21.0 ; platform_system == "Darwin"
torchaudio==2.6.0 ; platform_system == "Darwin"

# Windows (CPU)
torch==2.6.0 ; platform_system == "Windows"
torchvision==0.21.0 ; platform_system == "Windows"
torchaudio==2.6.0 ; platform_system == "Windows"

# Linux x86_64 (CUDA 12.4) — self-hosted GPU
torch==2.6.0+cu124 ; platform_system == "Linux" and platform_machine == "x86_64"
torchvision==0.21.0+cu124 ; platform_system == "Linux" and platform_machine == "x86_64"
torchaudio==2.6.0+cu124 ; platform_system == "Linux" and platform_machine == "x86_64"

# Core
fastapi==0.116.1
uvicorn==0.35.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
requests==2.32.5
psutil==7.0.0
python-multipart==0.0.20

# ML/Utils
transformers==4.56.1
sentencepiece==0.2.1
protobuf==3.20.3
pillow==11.0.0
soundfile==0.13.1
PyPDF2==3.0.1
Jinja2==3.1.4

reportlab==4.4.3
